pyproject.toml
[tool.poetry]
name = "demo-fastapi"
version = "0.1.0"
description = ""
authors = ["Jeter Chen <dali1756@gmail.com>"]
readme = "README.md"
packages = [{include = "apps"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
psycopg = "^3.2.3"
python-multipart = "^0.0.20"
jinja2 = "^3.1.5"
boto3 = "^1.35.92"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
